refactor: extract AdapterManager from main.ts (#197)
Move adapter-lifecycle concerns into a dedicated class:
- AdapterManager owns patchAdapter, restoreAdapter, replayOfflineQueue,
  showPendingEditsModal, fetchBinaryForBridge, stopResourceBridge,
  openOfflineQueue, makeThumbnailFetcherIfSupported,
  makeBinaryRangeFetcherIfSupported, and the PATCHED_METHODS constant.
- AdapterManager exposes a clean public surface:
    patch() → Promise<boolean>
    restore() → void
    dataAdapter: SftpDataAdapter | null (getter)
    isPatched() → boolean
    replayOfflineQueue(label) → Promise<void>
    showPendingEditsModal() → Promise<void>
- main.ts shrinks from 1225 → 807 lines; ConnectionManager (229) and
  AdapterManager (444) hold the two extracted domains.
- All 651 tests pass; tsc --noEmit clean.
